+++ Parser/parse_device.cpp
@@ -305,7 +305,7 @@ pointer gvsP_init_device (scheme *sc, pointer args)
 
     std::string idname = "unknown";
     if (!gvsParser->getParameter("id",idname)) {
-        appendNum(idname,gpTypeID.size());
+        appendNum(idname, (int)gpTypeID.size());
     }
     else if (gpTypeID.find(idname)!=gpTypeID.end()) {
         scheme_error("init-device: ID already assigned!");
