+++ Parser/parse_meshObjects.cpp
@@ -141,7 +141,7 @@ pointer gvsP_init_OBJmesh ( scheme *sc, pointer args )
 
     std::string idname = "unknown";
     if (!gvsParser->getParameter("id",idname)) {
-        appendNum(idname,gpTypeID.size());
+        appendNum(idname, (int)gpTypeID.size());
     }
     else if (gpTypeID.find(idname)!=gpTypeID.end()) {
         scheme_error("mesh-obj: ID is already assigned!");
